+++ TurboMinerVirtualTable.Unity/Assets/Scenes/Menu.unity
@@ -5488,7 +5488,7 @@ GameObject:
   - component: {fileID: 639578504}
   - component: {fileID: 639578503}
   m_Layer: 0
-  m_Name: MultiplayerManager
+  m_Name: HostService
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -5503,7 +5503,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 639578502}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c80721ef40bf142458a2728c523781a5, type: 3}
+  m_Script: {fileID: 11500000, guid: 41bde92c186a0e0469e01a7056ab21a4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   ServerPrefab: {fileID: 8115739811256896326, guid: ee7ed7196661e564c810f8b762eb7ff5,
@@ -5512,7 +5512,6 @@ MonoBehaviour:
     type: 3}
   PlayerNameInput: {fileID: 1580418766}
   HostIpInput: {fileID: 90387612}
-  StartButton: {fileID: 1139765679}
 --- !u!4 &639578504
 Transform:
   m_ObjectHideFlags: 0
@@ -5778,6 +5777,36 @@ Transform:
   m_Children: []
   m_Father: {fileID: 761556725}
   m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &757111675
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 757111676}
+  m_Layer: 0
+  m_Name: HostService
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &757111676
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 757111675}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 6
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &761556724
 GameObject:
