+++ TurboMinerVirtualTable.Unity/Assets/Scripts/Networking/ActionsQueue.cs
@@ -1,6 +1,9 @@
 ﻿using Assets.Scripts.Networking;
+using Assets.Scripts.Networking.Models;
+using Assets.Scripts.Settings.Models;
 using System;
 using System.Collections.Concurrent;
+using System.Linq;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
@@ -33,10 +36,10 @@ public class ActionsQueue : MonoBehaviour
             case MessageCommands.Server.Start:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.SetTilesSettings(message[1]);
-                    MultiplayerManager.Instance.SetCorridorsSettings(message[2]);
-                    MultiplayerManager.Instance.SetMapSizeSettings(message[3], message[4]);
-                    MultiplayerManager.Instance.SetPlayerSettings(message[5]);
+                    SetTilesSettings(message[1]);
+                    SetCorridorsSettings(message[2]);
+                    SetMapSizeSettings(message[3], message[4]);
+                    SetPlayerSettings(message[5]);
                     SceneManager.LoadScene("Table");
                 });
                 break;
@@ -82,58 +85,140 @@ public class ActionsQueue : MonoBehaviour
             case MessageCommands.Server.ElementPosition:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.SetElementsPositions(message[1]);
+                    SetElementsPositions(message[1]);
                 });
                 break;
 
             case MessageCommands.Server.ElementStopDrag:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.StopElementDrag(int.Parse(message[1]));
+                    StopElementDrag(int.Parse(message[1]));
                 });
                 break;
 
             case MessageCommands.Server.ElementLayer:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.IncrementElementsLayers(message[1]);
+                    IncrementElementsLayers(message[1]);
                 });
                 break;
 
             case MessageCommands.Server.ElementTurn:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.TurnElementOnOtherSide(int.Parse(message[1]));
+                    TurnElementOnOtherSide(int.Parse(message[1]));
                 });
                 break;
 
             case MessageCommands.Server.ElementRotate:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.RotateElement(int.Parse(message[1]));
+                    RotateElement(int.Parse(message[1]));
                 });
                 break;
 
             case MessageCommands.Server.ElementDestroy:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.DestroyElement(int.Parse(message[1]));
+                    DestroyElement(int.Parse(message[1]));
                 });
                 break;
 
             case MessageCommands.Server.RollDice:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.RollDice(int.Parse(message[1]));
+                    RollDice(int.Parse(message[1]));
                 });
                 break;
 
             case MessageCommands.Server.StackRefill:
                 mainThreadActions.Enqueue(() =>
                 {
-                    MultiplayerManager.Instance.RefillStack(message[1]);
+                    RefillStack(message[1]);
                 });
                 break;
         }
+    }
+
+    private void RefillStack(string stackRefillJson)
+    {
+        var stackRefill = JsonUtility.FromJson<StackRefill>(stackRefillJson);
+        var stack = Stack.Get(stackRefill.Id);
+        stack.Elements = stackRefill.RefillArray.ToList();
+        stack.SpawnOnTop();
+    }
+
+    private void DestroyElement(int elementId)
+    {
+        var element = Element.Get(elementId);
+        Destroy(element.gameObject);
+    }
+
+    private void RollDice(int finalDiceSide)
+    {
+        var dice = FindObjectOfType<Dice>();
+        dice.StartRolling(finalDiceSide);
+    }
+
+    private void IncrementElementsLayers(string jsonElementIds)
+    {
+        var elementIds = JsonUtility.FromJson<ElementIdArray>(jsonElementIds);
+        foreach (var elementId in elementIds.Array)
+        {
+            var element = Element.Get(elementId.Id);
+            element.IncrementLayerOrder();
+        }
+    }
+
+    private void TurnElementOnOtherSide(int elementId)
+    {
+        var element = Element.Get(elementId);
+        element.TurnOnOtherSide();
+    }
+
+    private void RotateElement(int elementId)
+    {
+        var element = Element.Get(elementId);
+        element.Rotate();
+    }
+
+    private void SetElementsPositions(string elementsJson)
+    {
+        var elements = JsonUtility.FromJson<ElementPositionArray>(elementsJson);
+        var containerElement = Element.Get(elements.Array[0].Id);
+        containerElement.IsDragged = true;
+        foreach (var elementPosition in elements.Array)
+        {
+            var element = Element.Get(elementPosition.Id);
+            element.transform.position = elementPosition.Position;
+        }
+    }
+
+    private void StopElementDrag(int elementId)
+    {
+        var element = Element.Get(elementId);
+        element.IsDragged = false;
+    }
+
+    private void SetTilesSettings(string tilesCsv)
+    {
+        var tiles = tilesCsv.Split(',');
+        GameSettings.Tiles = tiles.ToList();
+    }
+
+    private void SetCorridorsSettings(string corridorsCsv)
+    {
+        var corridors = corridorsCsv.Split(',');
+        GameSettings.Corridors = corridors.ToList();
+    }
+
+    private void SetMapSizeSettings(string width, string height)
+    {
+        GameSettings.MapSize = new MapSize(int.Parse(width), int.Parse(height));
+    }
+
+    private void SetPlayerSettings(string playerSettingsJson)
+    {
+        GameSettings.PlayersSettings = JsonUtility.FromJson<PlayerSettingsArray>(playerSettingsJson).Array;
     }
 }
